// final-project-team-tonder-fe/types/user-types.ts
export enum GENDER {
  MALE = 'male',
  FEMALE = 'female',
  OTHER = 'other',
}

export enum REASON {
  DATING = 'Muốn hẹn hò',
  FRIENDS = 'Muốn tâm sự',
  NEWRELATIONSHIP = 'Đang tìm một mối quan hệ mới',
}

export enum ALCOHOL {
  YES = 'Có uống',
  NEVER = 'Không bao giờ',
  SOMETIMES = 'Uống chút đỉnh',
  MORE = 'Đô bất tử',
}

export enum EDUCATION {
  NO = 'Không có',
  SECONDARY = 'Trung học',
  UNIVERSITY = 'Đại học',
  GRADUATED = 'Đã tốt nghiệp',
}